teddycrepineau/requirements/base.txt
astroid==2.1.0
Django==2.1.4
django-environ==0.4.5
djangorestframework==3.9.0
djangorestframework-simplejwt==3.3
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.6.1
PyJWT==1.7.1
pylint==2.2.2
pytz==2018.7
six==1.12.0
wrapt==1.10.11
